Fixed size_t type.

+++ src/Bpp/Seq/Container/SiteContainerTools.cpp
@@ -186,7 +186,7 @@ void SiteContainerTools::changeGapsToUnknownCharacters(SiteContainer& sites)
 {
   // NB: use iterators for a better algorithm?
   int unknownCode = sites.getAlphabet()->getUnknownCharacterCode();
-  for (unsigned int i = 0; i < sites.getNumberOfSites(); i++)
+  for (size_t i = 0; i < sites.getNumberOfSites(); i++)
   {
     for (unsigned int j = 0; j < sites.getNumberOfSequences(); j++)
     {
@@ -203,7 +203,7 @@ void SiteContainerTools::changeUnresolvedCharactersToGaps(SiteContainer& sites)
 {
   // NB: use iterators for a better algorithm?
   int gapCode = sites.getAlphabet()->getGapCharacterCode();
-  for (unsigned int i = 0; i < sites.getNumberOfSites(); i++)
+  for (size_t i = 0; i < sites.getNumberOfSites(); i++)
   {
     for (unsigned int j = 0; j < sites.getNumberOfSequences(); j++)
     {
@@ -221,7 +221,7 @@ SiteContainer* SiteContainerTools::removeGapOnlySites(const SiteContainer& sites
   vector<string> seqNames = sites.getSequencesNames();
   VectorSiteContainer* noGapCont = new VectorSiteContainer(seqNames.size(), sites.getAlphabet());
   noGapCont->setSequencesNames(seqNames, false);
-  for (unsigned int i = 0; i < sites.getNumberOfSites(); i++)
+  for (size_t i = 0; i < sites.getNumberOfSites(); i++)
   {
     const Site* site = &sites.getSite(i);
     if (!SiteTools::isGapOnly(*site))
@@ -268,7 +268,7 @@ SiteContainer* SiteContainerTools::removeGapOrUnresolvedOnlySites(const SiteCont
   vector<string> seqNames = sites.getSequencesNames();
   VectorSiteContainer* noGapCont = new VectorSiteContainer(seqNames.size(), sites.getAlphabet());
   noGapCont->setSequencesNames(seqNames, false);
-  for (unsigned int i = 0; i < sites.getNumberOfSites(); i++)
+  for (size_t i = 0; i < sites.getNumberOfSites(); i++)
   {
     const Site* site = &sites.getSite(i);
     if (!SiteTools::isGapOrUnresolvedOnly(*site))
@@ -315,7 +315,7 @@ SiteContainer* SiteContainerTools::removeGapSites(const SiteContainer& sites, do
   vector<string> seqNames = sites.getSequencesNames();
   VectorSiteContainer* noGapCont = new VectorSiteContainer(seqNames.size(), sites.getAlphabet());
   noGapCont->setSequencesNames(seqNames, false);
-  for (unsigned int i = 0; i < sites.getNumberOfSites(); ++i) {
+  for (size_t i = 0; i < sites.getNumberOfSites(); ++i) {
     map<int, double> freq;
     SiteTools::getFrequencies(sites.getSite(i), freq);
     if (freq[-1] <= maxFreqGaps)
